+++ epan/dissectors/packet-smb2.h
@@ -5,19 +5,7 @@
  * By Gerald Combs <gerald@wireshark.org>
  * Copyright 1998, 1999 Gerald Combs
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ * SPDX-License-Identifier: GPL-2.0-or-later
  */
 
 #ifndef __PACKET_SMB2_H__
@@ -81,7 +69,7 @@ typedef struct _smb2_saved_info_t {
 typedef struct _smb2_tid_info_t {
 	guint32 tid;
 	guint32 connect_frame;
-	guint16 share_type;
+	guint8 share_type;
 	char *name;
 } smb2_tid_info_t;
 
@@ -182,7 +170,7 @@ typedef struct _smb2_info_t {
 	guint32 status;
 	guint32 tid;
 	guint64 sesid;
-	gint64  msg_id;
+	guint64  msg_id;
 	guint32 flags;
 	smb2_eo_file_info_t	*eo_file_info; /* eo_smb extra info */
 	smb2_conv_info_t	*conv;
